+++ pkg/helm/content.go
@@ -83,10 +83,6 @@ func (h *HelmChartHandler) GetChart(c *gin.Context) {
 		return
 	}
 	if content.Metadata != nil && registry.IsOCI(repository.URL) {
-		// Synthesized OCI entries carry only name/version for most tags, so
-		// the archive's Chart.yaml fills in their display metadata. Classic
-		// repositories keep their index.yaml entries authoritative, matching
-		// ListCharts.
 		metadata := *content.Metadata
 		metadata.Name = entry.Name
 		metadata.Version = entry.Version
@@ -154,9 +150,7 @@ func (h *HelmChartHandler) GetChartContent(c *gin.Context) {
 	c.JSON(http.StatusOK, helmChartContentResponse{Templates: content.Templates})
 }
 
-// getChartEntry falls back to the newest entry when no version is requested
-// and IndexFile.Get finds no match: its "*" constraint skips prereleases, so a
-// prerelease-only chart would list in ListCharts yet 404 on detail lookups.
+// IndexFile.Get skips prereleases when no version is requested.
 func getChartEntry(indexFile *repo.IndexFile, chartName, version string) (*repo.ChartVersion, error) {
 	entry, err := indexFile.Get(chartName, version)
 	if err != nil && version == "" {
