+++ pkg/helmutil/chart.go
@@ -54,8 +54,7 @@ func LoadArchive(chartURL string, repository *model.HelmRepository) (*chart.Char
 
 	var registryClient *registry.Client
 	if parsedURL.Scheme == registry.OCIScheme && !OCIRefHasTagOrDigest(parsedURL) {
-		// Resolve the tag before the cache lookup so a tag-less URL never
-		// pins a stale "latest" for the TTL.
+		// Cache the resolved version instead of a mutable tag-less reference.
 		registryClient, err = newRegistryClient(repository, chartURL)
 		if err != nil {
 			return nil, err
@@ -130,48 +129,38 @@ func LoadArchive(chartURL string, repository *model.HelmRepository) (*chart.Char
 	return loadedChart, nil
 }
 
-// anonymousAuthorizer is shared by all credential-less registry clients: an
-// explicit anonymous authorizer keeps them anonymous — without it the SDK
-// falls back to the host's Helm/Docker credential stores, which leaks ambient
-// credentials and may invoke an interactive keychain helper. Sharing one
-// authorizer reuses its connections and cached auth tokens across calls.
-var anonymousAuthorizer = newAnonymousAuthorizer()
+var (
+	registryHTTPClient = &http.Client{
+		Transport: registry.NewTransport(false),
+		Timeout:   time.Duration(getter.DefaultHTTPTimeout) * time.Second,
+	}
+	anonymousAuthorizer = newAnonymousAuthorizer()
+)
 
 func newAnonymousAuthorizer() auth.Client {
 	authorizer := auth.Client{
-		Client: &http.Client{Transport: registry.NewTransport(false)},
+		Client: registryHTTPClient,
 		Cache:  auth.NewCache(),
 	}
 	authorizer.SetUserAgent("kite")
 	return authorizer
 }
 
-// newRegistryClient builds an OCI registry client, attaching the repository's
-// basic-auth credentials only when they belong to the target reference host.
 func newRegistryClient(repository *model.HelmRepository, targetURL string) (*registry.Client, error) {
+	options := []registry.ClientOption{registry.ClientOptHTTPClient(registryHTTPClient)}
 	if repositoryCredentialsApply(repository, targetURL) {
-		return registry.NewClient(registry.ClientOptBasicAuth(repository.Username, string(repository.Password)))
+		options = append(options, registry.ClientOptBasicAuth(repository.Username, string(repository.Password)))
+	} else {
+		options = append(options, registry.ClientOptAuthorizer(anonymousAuthorizer))
 	}
-	return registry.NewClient(registry.ClientOptAuthorizer(anonymousAuthorizer))
+	return registry.NewClient(options...)
 }
 
 func repositoryCredentialsApply(repository *model.HelmRepository, targetURL string) bool {
 	return repository != nil && repository.Username != "" && sameURLHost(repository.URL, targetURL)
 }
 
-// ListOCIChartVersions returns the semver tags of the chart at an oci:// URL,
-// newest first. Non-semver tags (e.g. "latest", "v1.2.3") are filtered out by
-// the registry client.
-func ListOCIChartVersions(repository *model.HelmRepository, chartURL string) ([]string, error) {
-	registryClient, err := newRegistryClient(repository, chartURL)
-	if err != nil {
-		return nil, err
-	}
-	return registryClient.Tags(ociReference(chartURL))
-}
-
-// OCIChartName returns the chart name an oci:// repository URL points at: the
-// last segment of the registry path.
+// OCIChartName returns the last path segment of an OCI repository URL.
 func OCIChartName(repositoryURL string) (string, error) {
 	parsedURL, err := url.Parse(repositoryURL)
 	if err != nil {
@@ -184,16 +173,8 @@ func OCIChartName(repositoryURL string) (string, error) {
 	return name, nil
 }
 
-// OCIRefHasTagOrDigest reports whether an oci:// URL pins a tag or digest.
 func OCIRefHasTagOrDigest(parsedURL *url.URL) bool {
 	return strings.Contains(path.Base(parsedURL.Path), ":") || strings.Contains(parsedURL.Path, "@")
-}
-
-// OCIChartVersionURL builds the chart URL for one registry tag. The
-// "<repository URL>:<tag>" format is load-bearing for
-// MatchesRepositoryCacheKey.
-func OCIChartVersionURL(repositoryURL, tag string) string {
-	return strings.TrimRight(repositoryURL, "/") + ":" + tag
 }
 
 func ociReference(chartURL string) string {
@@ -227,11 +208,7 @@ func archiveCacheKey(chartURL string) string {
 	return chartURL
 }
 
-// MatchesRepositoryCacheKey reports whether a cached chart URL belongs to the
-// repository: an exact match, a path under it, or — only for oci://
-// repositories, whose archives are cached under "<repository URL>:<tag>" — a
-// tag reference. The tag form is not matched for http repositories so they
-// cannot evict a sibling on the same host that carries an explicit port.
+// MatchesRepositoryCacheKey matches paths and OCI tag references under a repository URL.
 func MatchesRepositoryCacheKey(repositoryURL, key string) bool {
 	if key == repositoryURL {
 		return true
